refactor(activation): remove dead contract fallbacks
Constraint: select only the canary kind already authenticated by activation input verification.\nRejected: retaining unreachable service and operation inference as misleading defensive code.\nConfidence: high; targeted activation and rollout tests plus vet pass.\nScope-risk: low; removes unreachable branches, corrects one legacy fixture, and aligns test integer types.\nNot-tested: no new remote qualification because qualified adapter bytes and runtime behavior are unchanged.

+++ internal/rolloutdriver/verify_test.go
@@ -463,8 +463,8 @@ func rolloutOpenClawTerminal(t *testing.T, activationID string) []byte {
 			DurationMS   int64    `json:"duration_ms"`
 			Model        string   `json:"model"`
 			Provider     string   `json:"provider"`
-			ToolCalls    int64    `json:"tool_calls"`
-			ToolFailures int64    `json:"tool_failures"`
+			ToolCalls    int      `json:"tool_calls"`
+			ToolFailures int      `json:"tool_failures"`
 			Tools        []string `json:"tools"`
 		} `json:"meta"`
 	}{}
